feat: add icon support, revise asset keys
