+++ src/index.ts
@@ -44,8 +44,9 @@ export { InterstitialAd } from './ads/InterstitialAd';
 export { RewardedAd } from './ads/RewardedAd';
 export { RewardedInterstitialAd } from './ads/RewardedInterstitialAd';
 export { BannerAd } from './ads/BannerAd';
-export { NativeAd } from './ads/NativeAd';
-export { NativeAdView, NativeAsset } from './ads/NativeAdView';
+export { NativeAd } from './ads/native-ad/NativeAd';
+export { NativeAdView } from './ads/native-ad/NativeAdView';
+export { NativeAsset, NativeAssetType } from './ads/native-ad/NativeAsset';
 export { GAMBannerAd } from './ads/GAMBannerAd';
 export { GAMInterstitialAd } from './ads/GAMInterstitialAd';
 export { useAppOpenAd } from './hooks/useAppOpenAd';
