+++ src/specs/modules/NativeGoogleMobileAdsNativeModule.ts
@@ -19,18 +19,6 @@ import type { TurboModule } from 'react-native';
 import { TurboModuleRegistry } from 'react-native';
 import type { Double, UnsafeObject } from 'react-native/Libraries/Types/CodegenTypes';
 
-export enum NativeAssetType {
-  ADVERTISER = 'advertiser',
-  BODY = 'body',
-  CTA = 'callToAction',
-  HEADLINE = 'headline',
-  PRICE = 'price',
-  STORE = 'store',
-  RATINGS = 'ratings',
-  ICONS = 'icons',
-  IMAGES = 'images',
-}
-
 export type NativeAdProps = {
   responseId: string;
   advertiser: string | null;
@@ -39,7 +27,7 @@ export type NativeAdProps = {
   headline: string | null;
   price: string | null;
   store: string | null;
-  ratings: Double | null;
+  starRating: Double | null;
   icon: NativeAdImage | null;
   images: Array<NativeAdImage> | null;
   extras: UnsafeObject | null;
